Fix validate script not spawning tasks correctly
We are spawning the tasks with sh -c which expects a string of the full
command. The old implementation was splitting the command into an args
array, meaning that sh -c was only looking at the first argument. For,
e.g. jshint, this means that instead of "jshint ." it was running
"jshint". Which is basically a no-op that will always pass.

Fixes #17