Support Git submodules.
When the Git repo being installed or the current working directory is a Git
submodule inside another Git repo, be aware of it and treat it as a self-contained
Git repo, ignoring the container repo entirely.

Essentially, this means that we don't assume there is a ".git" directory in this repo.
If we find a ".git" _file_, this means it's a submodule, so we parse the single-line
file which indicates the path to its ".git"-ish directory, tucked away somewhere inside
the container repo.  Then everything works as expected.